Depatch 1.18.8
* Added working patch for BP update reduce.
and option to exclude modded or vanilla PB block by SubSype
All changed can be done LIVE!!!

Some cleanup.

+++ DePatch/GamePatches/MyCameraBlockPatch.cs
@@ -11,7 +11,7 @@ namespace DePatch
             if (!DePatchPlugin.Instance.Config.Enabled)
                 return;
 
-            __instance.BlockDefinition.RaycastDistanceLimit = (double)DePatchPlugin.Instance.Config.RaycastLimit;
+            __instance.BlockDefinition.RaycastDistanceLimit = DePatchPlugin.Instance.Config.RaycastLimit;
         }
     }
 }
